refactor: move subscription and acknowledgment interfaces to package-specific modules
