+++ go/ack/interface.go
@@ -1,4 +1,4 @@
-package interfaces
+package ack
 
 import (
 	"context"
